Moved the dataControl controller out of the channelsModal controller
To manage the controller in a seperate file

OHM-656
