+++ app/scripts/controllers/index.js
@@ -41,3 +41,4 @@ export * from './channel-tabs/basicInfo'
 export * from './channel-tabs/requestMatching'
 export * from './channel-tabs/userAccess'
 export * from './channel-tabs/routes'
+export * from './channel-tabs/dataControl'
